feat: add zipformer

+++ examples/vad_whisper.rs
@@ -79,7 +79,7 @@ fn main() -> Result<()> {
                 let segment = vad.front();
                 let start_sec = (segment.start as f32) / sample_rate as f32;
                 let duration_sec = (segment.samples.len() as f32) / sample_rate as f32;
-                let transcript = recognizer.transcribe(sample_rate, segment.samples.clone());
+                let transcript = recognizer.transcribe(sample_rate as u32, segment.samples.clone());
 
                 // Compute the speaker embedding
                 let mut embedding =
@@ -114,7 +114,7 @@ fn main() -> Result<()> {
             let segment = vad.front();
             let start_sec = (segment.start as f32) / sample_rate as f32;
             let duration_sec = (segment.samples.len() as f32) / sample_rate as f32;
-            let transcript = recognizer.transcribe(sample_rate, segment.samples.clone());
+            let transcript = recognizer.transcribe(sample_rate as u32, segment.samples.clone());
 
             // Compute the speaker embedding
             let mut embedding =
